+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](semver).
 ## [Unreleased]
 
 ### Added
+- Admin dashboard data sorting. [#99]
+- Ability to download sorted data. [#129]
 
 ### Changed